chatbot_api/requirements.txt
asyncio==3.4.3
fastapi==0.115.0
langchain==0.3.1
langchain-chroma==0.1.4
langchain-community==0.3.1
langchain-openai==0.2.1
langchainhub==0.1.21
neo4j==5.25.0
numpy==1.26.4
pydantic==2.9.2
python-dotenv==1.0.1
pypdf==5.0.0
opentelemetry-api==1.27.0
uvicorn==0.31.0